// ts/features/itwallet/presentation/proximity/analytics/enum.ts
export enum ITW_PROXIMITY_SCREENVIEW_EVENTS {
  ITW_BLUETOOTH_ACCESS = "ITW_BLUETOOTH_ACCESS",
  ITW_BLUETOOTH_ACCESS_DENIED = "ITW_BLUETOOTH_ACCESS_DENIED",
  ITW_BLUETOOTH_BLOCK = "ITW_BLUETOOTH_BLOCK",
  ITW_BLUETOOTH_ACTIVATION = "ITW_BLUETOOTH_ACTIVATION",
  ITW_BLUETOOTH_NOT_ACTIVATED = "ITW_BLUETOOTH_NOT_ACTIVATED",
  ITW_PROXIMITY_QR_CODE = "ITW_PROXIMITY_QR_CODE",
  ITW_PROXIMITY_DATA_SHARE = "ITW_PROXIMITY_DATA_SHARE",
  ITW_PROXIMITY_UX_SUCCESS = "ITW_PROXIMITY_UX_SUCCESS",
  ITW_QR_CODE = "ITW_QR_CODE"
}

export enum ITW_PROXIMITY_ACTIONS_EVENTS {
  ITW_PROXIMITY_SHOW_QR_CODE = "ITW_PROXIMITY_SHOW_QR_CODE",
  ITW_BLUETOOTH_ACCESS_CLOSE = "ITW_BLUETOOTH_ACCESS_CLOSE",
  ITW_BLUETOOTH_ACCESS_GO_TO_SETTINGS = "ITW_BLUETOOTH_ACCESS_GO_TO_SETTINGS",
  ITW_BLUETOOTH_BLOCK_ACTION = "ITW_BLUETOOTH_BLOCK_ACTION",
  ITW_BLUETOOTH_ACTIVATION_CLOSE = "ITW_BLUETOOTH_ACTIVATION_CLOSE",
  ITW_BLUETOOTH_ACTIVATION_GO_TO_SETTINGS = "ITW_BLUETOOTH_ACTIVATION_GO_TO_SETTINGS",
  ITW_PROXIMITY_QR_CODE_LOADING_RETRY = "ITW_PROXIMITY_QR_CODE_LOADING_RETRY",
  ITW_PROXIMITY_UX_CONVERSION = "ITW_PROXIMITY_UX_CONVERSION",
  ITW_PROXIMITY_UNOFFICIAL_VERIFIER_BOTTOMSHEET = "ITW_PROXIMITY_UNOFFICIAL_VERIFIER_BOTTOMSHEET",
  ITW_SHOW_QR_CODE = "ITW_SHOW_QR_CODE",
  ITW_START_REISSUING_PID = "ITW_START_REISSUING_PID",
  ITW_QR_CODE_LOADING_RETRY = "ITW_QR_CODE_LOADING_RETRY"
}

export enum ITW_PROXIMITY_ERRORS_EVENTS {
  ITW_PROXIMITY_QR_CODE_LOADING_FAILURE = "ITW_PROXIMITY_QR_CODE_LOADING_FAILURE",
  ITW_PROXIMITY_GENERIC_ERROR = "ITW_PROXIMITY_GENERIC_ERROR",
  ITW_PROXIMITY_TIMEOUT = "ITW_PROXIMITY_TIMEOUT",
  ITW_PROXIMITY_UNEXPECTED_FAILURE = "ITW_PROXIMITY_UNEXPECTED_FAILURE",
  ITW_PROXIMITY_UNOFFICIAL_VERIFIER = "ITW_PROXIMITY_UNOFFICIAL_VERIFIER",
  ITW_QR_CODE_LOADING_FAILURE = "ITW_QR_CODE_LOADING_FAILURE"
}